requirements.txt
Django==1.9.2
django-autoslug==1.9.3
django-crispy-forms==1.6.0
django-markdown==0.8.4
django-registration-redux==1.4
Markdown==2.6.5
wheel==0.29.0
